Extract a few utility methods. Use a single WebDriver instance instead of creating a new one for every screenshot.
